Incorporate mercury navigation (it wasnt mercury...) to discussions navigation tests
