+++ src/test/java/com/wikia/webdriver/testsuites/testSuite.xml
@@ -155,7 +155,6 @@
             <class name="com.wikia.webdriver.testcases.mediatests.VetAddingVideoTests"/>
             <class name="com.wikia.webdriver.testcases.mediatests.videosmodule.VideosModuleTests"/>
             <class name="com.wikia.webdriver.testcases.mercurytests.PortableInfoboxTests"/>
-            <class name="com.wikia.webdriver.testcases.discussions.NavigationTests"/>
             <class name="com.wikia.webdriver.testcases.messagewall.MessageWallFeaturesTests"/>
             <class name="com.wikia.webdriver.testcases.messagewall.MessageWallTests"/>
             <class name="com.wikia.webdriver.testcases.mobilewikitests.ArticlePageTests"/>
